feat: replace old login screen with new Google login UI
